clean a bit
